v0.15
Implement the lexical rules for the chosen programming language, in the  lexer.jflex file, from the JavaCup project.
The rules must include the regular expressions needed to identify the code elements as well as the mapping between the language specific elements (keywords, operators, ...) and the java code that creates the associated symbols.